created initial backend tests
