+++ .gitignore
@@ -65,3 +65,6 @@ typings/
 
 # next.js build output
 .next
+
+# sqlite files (used during testing)
+*.sqlite
