+++ backend.code-workspace
@@ -23,6 +23,7 @@
       ".eslintignore": "ignore",
       ".flowconfig": "ignore",
       ".npmrc": "ignore",
+      "testenv": "ignore",
       ".yarnrc": "ignore"
     },
