+++ frontend.code-workspace
@@ -22,6 +22,7 @@
       ".eslintignore": "ignore",
       ".flowconfig": "ignore",
       ".npmrc": "ignore",
+      "testenv": "ignore",
       ".yarnrc": "ignore"
     },
