+++ packages/backend/.env.example
@@ -4,7 +4,7 @@ DB_HOST=sample_host
 DB_USERNAME=sample_username
 DB_PASSWORD=sample_password
 DB_DATABASE=sample_database_name
-
+DB_TYPE=mysql
 
 # For authentication security
 ## The type of hashing used, see https://nodejs.org/api/crypto.html#crypto_crypto_createhmac_algorithm_key_options
