+++ packages/backend/test/helpers/loadTestEnv.js
@@ -1,0 +1,5 @@
+import { config as loadEnv } from 'dotenv';
+
+loadEnv({
+  path: require.resolve('../testenv'),
+});
